Move README.md check to the correct place
This is because it should be checked first whether I *contributed* to a repository, rather than if README.md does not exist.

I would consider my own project unfinished if README.md does not exist, but for other people's stuff, I really could care less, that just means I contributed to it.